staging/pkg/mod/golang.org/x/net@v0.0.0-20210520170846-37e1c6afe023/html/testdata/webkit/tests16.dat

#data
<!doctype html><script>
#errors
(1,23): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|   <body>

#data
<!doctype html><script>a
#errors
(1,24): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "a"
|   <body>

#data
<!doctype html><script><
#errors
(1,24): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<"
|   <body>

#data
<!doctype html><script></
#errors
(1,25): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</"
|   <body>

#data
<!doctype html><script></S
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</S"
|   <body>

#data
<!doctype html><script></SC
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SC"
|   <body>

#data
<!doctype html><script></SCR
#errors
(1,28): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCR"
|   <body>

#data
<!doctype html><script></SCRI
#errors
(1,29): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCRI"
|   <body>

#data
<!doctype html><script></SCRIP
#errors
(1,30): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCRIP"
|   <body>

#data
<!doctype html><script></SCRIPT
#errors
(1,31): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCRIPT"
|   <body>

#data
<!doctype html><script></SCRIPT 
#errors
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|   <body>

#data
<!doctype html><script></s
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</s"
|   <body>

#data
<!doctype html><script></sc
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</sc"
|   <body>

#data
<!doctype html><script></scr
#errors
(1,28): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</scr"
|   <body>

#data
<!doctype html><script></scri
#errors
(1,29): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</scri"
|   <body>

#data
<!doctype html><script></scrip
#errors
(1,30): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</scrip"
|   <body>

#data
<!doctype html><script></script
#errors
(1,31): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</script"
|   <body>

#data
<!doctype html><script></script 
#errors
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|   <body>

#data
<!doctype html><script><!
#errors
(1,25): expected-script-data-but-got-eof
(1,25): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!"
|   <body>

#data
<!doctype html><script><!a
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!a"
|   <body>

#data
<!doctype html><script><!-
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!-"
|   <body>

#data
<!doctype html><script><!-a
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!-a"
|   <body>

#data
<!doctype html><script><!--
#errors
(1,27): expected-named-closing-tag-but-got-eof
(1,27): unexpected-eof-in-text-mode
#new-errors
(1:28) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<!doctype html><script><!--a
#errors
(1,28): expected-named-closing-tag-but-got-eof
(1,28): unexpected-eof-in-text-mode
#new-errors
(1:29) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--a"
|   <body>

#data
<!doctype html><script><!--<
#errors
(1,28): expected-named-closing-tag-but-got-eof
(1,28): unexpected-eof-in-text-mode
#new-errors
(1:29) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<"
|   <body>

#data
<!doctype html><script><!--<a
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<a"
|   <body>

#data
<!doctype html><script><!--</
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--</"
|   <body>

#data
<!doctype html><script><!--</script
#errors
(1,35): expected-named-closing-tag-but-got-eof
(1,35): unexpected-eof-in-text-mode
#new-errors
(1:36) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--</script"
|   <body>

#data
<!doctype html><script><!--</script 
#errors
(1,36): expected-attribute-name-but-got-eof
(1,36): expected-named-closing-tag-but-got-eof
#new-errors
(1:37) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<!doctype html><script><!--<s
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<s"
|   <body>

#data
<!doctype html><script><!--<script
#errors
(1,34): expected-named-closing-tag-but-got-eof
(1,34): unexpected-eof-in-text-mode
#new-errors
(1:35) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script"
|   <body>

#data
<!doctype html><script><!--<script 
#errors
(1,35): eof-in-script-in-script
(1,35): expected-named-closing-tag-but-got-eof
#new-errors
(1:36) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script "
|   <body>

#data
<!doctype html><script><!--<script <
#errors
(1,36): eof-in-script-in-script
(1,36): expected-named-closing-tag-but-got-eof
#new-errors
(1:37) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script <"
|   <body>

#data
<!doctype html><script><!--<script <a
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script <a"
|   <body>

#data
<!doctype html><script><!--<script </
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </"
|   <body>

#data
<!doctype html><script><!--<script </s
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </s"
|   <body>

#data
<!doctype html><script><!--<script </script
#errors
(1,43): eof-in-script-in-script
(1,43): expected-named-closing-tag-but-got-eof
#new-errors
(1:44) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script"
|   <body>

#data
<!doctype html><script><!--<script </scripta
#errors
(1,44): eof-in-script-in-script
(1,44): expected-named-closing-tag-but-got-eof
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </scripta"
|   <body>

#data
<!doctype html><script><!--<script </script 
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script </script>
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script>"
|   <body>

#data
<!doctype html><script><!--<script </script/
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#new-errors
(1:45) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script/"
|   <body>

#data
<!doctype html><script><!--<script </script <
#errors
(1,45): expected-named-closing-tag-but-got-eof
(1,45): unexpected-eof-in-text-mode
#new-errors
(1:46) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script <"
|   <body>

#data
<!doctype html><script><!--<script </script <a
#errors
(1,46): expected-named-closing-tag-but-got-eof
(1,46): unexpected-eof-in-text-mode
#new-errors
(1:47) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script <a"
|   <body>

#data
<!doctype html><script><!--<script </script </
#errors
(1,46): expected-named-closing-tag-but-got-eof
(1,46): unexpected-eof-in-text-mode
#new-errors
(1:47) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script </"
|   <body>

#data
<!doctype html><script><!--<script </script </script
#errors
(1,52): expected-named-closing-tag-but-got-eof
(1,52): unexpected-eof-in-text-mode
#new-errors
(1:53) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script </script"
|   <body>

#data
<!doctype html><script><!--<script </script </script 
#errors
(1,53): expected-attribute-name-but-got-eof
(1,53): expected-named-closing-tag-but-got-eof
#new-errors
(1:54) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script </script </script/
#errors
(1,53): unexpected-EOF-after-solidus-in-tag
(1,53): expected-named-closing-tag-but-got-eof
#new-errors
(1:54) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script </script </script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script -
#errors
(1,36): eof-in-script-in-script
(1,36): expected-named-closing-tag-but-got-eof
#new-errors
(1:37) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -"
|   <body>

#data
<!doctype html><script><!--<script -a
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -a"
|   <body>

#data
<!doctype html><script><!--<script -<
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -<"
|   <body>

#data
<!doctype html><script><!--<script --
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --"
|   <body>

#data
<!doctype html><script><!--<script --a
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --a"
|   <body>

#data
<!doctype html><script><!--<script --<
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --<"
|   <body>

#data
<!doctype html><script><!--<script -->
#errors
(1,38): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script --><
#errors
(1,39): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --><"
|   <body>

#data
<!doctype html><script><!--<script --></
#errors
(1,40): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --></"
|   <body>

#data
<!doctype html><script><!--<script --></script
#errors
(1,46): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --></script"
|   <body>

#data
<!doctype html><script><!--<script --></script 
#errors
(1,47): expected-attribute-name-but-got-eof
(1,47): expected-named-closing-tag-but-got-eof
#new-errors
(1:48) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script --></script/
#errors
(1,47): unexpected-EOF-after-solidus-in-tag
(1,47): expected-named-closing-tag-but-got-eof
#new-errors
(1:48) eof-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script --></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script><\/script>--></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script><\/script>-->"
|   <body>

#data
<!doctype html><script><!--<script></scr'+'ipt>--></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt>-->"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>--><!--</script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>--><!--"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>-- ></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>-- >"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>- -></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- ->"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>- - ></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- - >"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>-></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>->"
|   <body>

#data
<!doctype html><script><!--<script>--!></script>X
#errors
(1,49): expected-named-closing-tag-but-got-eof
(1,49): unexpected-EOF-in-text-mode
#new-errors
(1:50) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script>--!></script>X"
|   <body>

#data
<!doctype html><script><!--<scr'+'ipt></script>--></script>
#errors
(1,59): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<scr'+'ipt>"
|   <body>
|     "-->"

#data
<!doctype html><script><!--<script></scr'+'ipt></script>X
#errors
(1,57): expected-named-closing-tag-but-got-eof
(1,57): unexpected-eof-in-text-mode
#new-errors
(1:58) eof-in-script-html-comment-like-text
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt></script>X"
|   <body>

#data
<!doctype html><style><!--<style></style>--></style>
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--<style>"
|   <body>
|     "-->"

#data
<!doctype html><style><!--</style>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--"
|   <body>
|     "X"

#data
<!doctype html><style><!--...</style>...--></style>
#errors
(1,51): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--..."
|   <body>
|     "...-->"

#data
<!doctype html><style><!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style></style>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style>"
|   <body>
|     "X"

#data
<!doctype html><style><!--...<style><!--...--!></style>--></style>
#errors
(1,66): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--...<style><!--...--!>"
|   <body>
|     "-->"

#data
<!doctype html><style><!--...</style><!-- --><style>@import ...</style>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--..."
|     <!--   -->
|     <style>
|       "@import ..."
|   <body>

#data
<!doctype html><style>...<style><!--...</style><!-- --></style>
#errors
(1,63): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "...<style><!--..."
|     <!--   -->
|   <body>

#data
<!doctype html><style>...<!--[if IE]><style>...</style>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "...<!--[if IE]><style>..."
|   <body>
|     "X"

#data
<!doctype html><title><!--<title></title>--></title>
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "<!--<title>"
|   <body>
|     "-->"

#data
<!doctype html><title>&lt;/title></title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "</title>"
|   <body>

#data
<!doctype html><title>foo/title><link></head><body>X
#errors
(1,52): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "foo/title><link></head><body>X"
|   <body>

#data
<!doctype html><noscript><!--<noscript></noscript>--></noscript>
#errors
(1,64): unexpected-end-tag
#script-on
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<!--<noscript>"
|   <body>
|     "-->"

#data
<!doctype html><noscript><!--<noscript></noscript>--></noscript>
#errors
#script-off
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       <!-- <noscript></noscript> -->
|   <body>

#data
<!doctype html><noscript><!--</noscript>X<noscript>--></noscript>
#errors
#script-on
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<!--"
|   <body>
|     "X"
|     <noscript>
|       "-->"

#data
<!doctype html><noscript><!--</noscript>X<noscript>--></noscript>
#errors
#script-off
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       <!-- </noscript>X<noscript> -->
|   <body>

#data
<!doctype html><noscript><iframe></noscript>X
#errors
#script-on
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<iframe>"
|   <body>
|     "X"

#data
<!doctype html><noscript><iframe></noscript>X
#errors
 * (1,34) unexpected token in head noscript
 * (1,46) unexpected EOF
#script-off
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <iframe>
|       "</noscript>X"

#data
<!doctype html><noframes><!--<noframes></noframes>--></noframes>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noframes>
|       "<!--<noframes>"
|   <body>
|     "-->"

#data
<!doctype html><noframes><body><script><!--...</script></body></noframes></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noframes>
|       "<body><script><!--...</script></body>"
|   <body>

#data
<!doctype html><textarea><!--<textarea></textarea>--></textarea>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!--<textarea>"
|     "-->"

#data
<!doctype html><textarea>&lt;/textarea></textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "</textarea>"

#data
<!doctype html><textarea>&lt;</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<"

#data
<!doctype html><textarea>a&lt;b</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "a<b"

#data
<!doctype html><iframe><!--<iframe></iframe>--></iframe>
#errors
(1,56): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "<!--<iframe>"
|     "-->"

#data
<!doctype html><iframe>...<!--X->...<!--/X->...</iframe>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "...<!--X->...<!--/X->..."

#data
<!doctype html><xmp><!--<xmp></xmp>--></xmp>
#errors
(1,44): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <xmp>
|       "<!--<xmp>"
|     "-->"

#data
<!doctype html><noembed><!--<noembed></noembed>--></noembed>
#errors
(1,60): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <noembed>
|       "<!--<noembed>"
|     "-->"

#data
<script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,8): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<script>a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,9): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "a"
|   <body>

#data
<script><
#errors
(1,8): expected-doctype-but-got-start-tag
(1,9): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<"
|   <body>

#data
<script></
#errors
(1,8): expected-doctype-but-got-start-tag
(1,10): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</"
|   <body>

#data
<script></S
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</S"
|   <body>

#data
<script></SC
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SC"
|   <body>

#data
<script></SCR
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCR"
|   <body>

#data
<script></SCRI
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCRI"
|   <body>

#data
<script></SCRIP
#errors
(1,8): expected-doctype-but-got-start-tag
(1,15): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCRIP"
|   <body>

#data
<script></SCRIPT
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCRIPT"
|   <body>

#data
<script></SCRIPT 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,17): expected-attribute-name-but-got-eof
(1,17): expected-named-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<script></s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</s"
|   <body>

#data
<script></sc
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</sc"
|   <body>

#data
<script></scr
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</scr"
|   <body>

#data
<script></scri
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</scri"
|   <body>

#data
<script></scrip
#errors
(1,8): expected-doctype-but-got-start-tag
(1,15): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</scrip"
|   <body>

#data
<script></script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</script"
|   <body>

#data
<script></script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,17): expected-attribute-name-but-got-eof
(1,17): expected-named-closing-tag-but-got-eof
#new-errors
(1:18) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<script><!
#errors
(1,8): expected-doctype-but-got-start-tag
(1,10): expected-script-data-but-got-eof
(1,10): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!"
|   <body>

#data
<script><!a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!a"
|   <body>

#data
<script><!-
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!-"
|   <body>

#data
<script><!-a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!-a"
|   <body>

#data
<script><!--
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
(1,12): unexpected-eof-in-text-mode
#new-errors
(1:13) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<script><!--a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
(1,13): unexpected-eof-in-text-mode
#new-errors
(1:14) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--a"
|   <body>

#data
<script><!--<
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
(1,13): unexpected-eof-in-text-mode
#new-errors
(1:14) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<"
|   <body>

#data
<script><!--<a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#new-errors
(1:15) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<a"
|   <body>

#data
<script><!--</
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#new-errors
(1:15) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--</"
|   <body>

#data
<script><!--</script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,20): expected-named-closing-tag-but-got-eof
(1,20): unexpected-eof-in-text-mode
#new-errors
(1:21) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--</script"
|   <body>

#data
<script><!--</script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): expected-attribute-name-but-got-eof
(1,21): expected-named-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<script><!--<s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#new-errors
(1:15) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<s"
|   <body>

#data
<script><!--<script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,19): expected-named-closing-tag-but-got-eof
(1,19): unexpected-eof-in-text-mode
#new-errors
(1:20) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script"
|   <body>

#data
<script><!--<script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,20): eof-in-script-in-script
(1,20): expected-named-closing-tag-but-got-eof
#new-errors
(1:21) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script "
|   <body>

#data
<script><!--<script <
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): eof-in-script-in-script
(1,21): expected-named-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script <"
|   <body>

#data
<script><!--<script <a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script <a"
|   <body>

#data
<script><!--<script </
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </"
|   <body>

#data
<script><!--<script </s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): eof-in-script-in-script
(1,23): expected-named-closing-tag-but-got-eof
#new-errors
(1:24) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </s"
|   <body>

#data
<script><!--<script </script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,28): eof-in-script-in-script
(1,28): expected-named-closing-tag-but-got-eof
#new-errors
(1:29) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script"
|   <body>

#data
<script><!--<script </scripta
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): eof-in-script-in-script
(1,29): expected-named-closing-tag-but-got-eof
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </scripta"
|   <body>

#data
<script><!--<script </script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script </script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script>"
|   <body>

#data
<script><!--<script </script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#new-errors
(1:30) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script/"
|   <body>

#data
<script><!--<script </script <
#errors
(1,8): expected-doctype-but-got-start-tag
(1,30): expected-named-closing-tag-but-got-eof
(1,30): unexpected-eof-in-text-mode
#new-errors
(1:31) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script <"
|   <body>

#data
<script><!--<script </script <a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
(1,31): unexpected-eof-in-text-mode
#new-errors
(1:32) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script <a"
|   <body>

#data
<script><!--<script </script </
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
(1,31): unexpected-eof-in-text-mode
#new-errors
(1:32) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script </"
|   <body>

#data
<script><!--<script </script </script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,37): expected-named-closing-tag-but-got-eof
(1,37): unexpected-eof-in-text-mode
#new-errors
(1:38) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script </script"
|   <body>

#data
<script><!--<script </script </script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,38): expected-attribute-name-but-got-eof
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script </script </script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,38): unexpected-EOF-after-solidus-in-tag
(1,38): expected-named-closing-tag-but-got-eof
#new-errors
(1:39) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script </script </script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script -
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): eof-in-script-in-script
(1,21): expected-named-closing-tag-but-got-eof
#new-errors
(1:22) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -"
|   <body>

#data
<script><!--<script -a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -a"
|   <body>

#data
<script><!--<script --
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#new-errors
(1:23) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --"
|   <body>

#data
<script><!--<script --a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): eof-in-script-in-script
(1,23): expected-named-closing-tag-but-got-eof
#new-errors
(1:24) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --a"
|   <body>

#data
<script><!--<script -->
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script --><
#errors
(1,8): expected-doctype-but-got-start-tag
(1,24): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --><"
|   <body>

#data
<script><!--<script --></
#errors
(1,8): expected-doctype-but-got-start-tag
(1,25): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --></"
|   <body>

#data
<script><!--<script --></script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --></script"
|   <body>

#data
<script><!--<script --></script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script --></script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,32): unexpected-EOF-after-solidus-in-tag
(1,32): expected-named-closing-tag-but-got-eof
#new-errors
(1:33) eof-in-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script --></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script><\/script>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script><\/script>-->"
|   <body>

#data
<script><!--<script></scr'+'ipt>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt>-->"
|   <body>

#data
<script><!--<script></script><script></script></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>"
|   <body>

#data
<script><!--<script></script><script></script>--><!--</script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>--><!--"
|   <body>

#data
<script><!--<script></script><script></script>-- ></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>-- >"
|   <body>

#data
<script><!--<script></script><script></script>- -></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- ->"
|   <body>

#data
<script><!--<script></script><script></script>- - ></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- - >"
|   <body>

#data
<script><!--<script></script><script></script>-></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>->"
|   <body>

#data
<script><!--<script>--!></script>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,34): expected-named-closing-tag-but-got-eof
(1,34): unexpected-eof-in-text-mode
#new-errors
(1:35) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script>--!></script>X"
|   <body>

#data
<script><!--<scr'+'ipt></script>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,44): unexpected-end-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<scr'+'ipt>"
|   <body>
|     "-->"

#data
<script><!--<script></scr'+'ipt></script>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,42): expected-named-closing-tag-but-got-eof
(1,42): unexpected-eof-in-text-mode
#new-errors
(1:43) eof-in-script-html-comment-like-text
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt></script>X"
|   <body>

#data
<style><!--<style></style>--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--<style>"
|   <body>
|     "-->"

#data
<style><!--</style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--"
|   <body>
|     "X"

#data
<style><!--...</style>...--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,36): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--..."
|   <body>
|     "...-->"

#data
<style><!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style></style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style>"
|   <body>
|     "X"

#data
<style><!--...<style><!--...--!></style>--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,51): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--...<style><!--...--!>"
|   <body>
|     "-->"

#data
<style><!--...</style><!-- --><style>@import ...</style>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--..."
|     <!--   -->
|     <style>
|       "@import ..."
|   <body>

#data
<style>...<style><!--...</style><!-- --></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,48): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "...<style><!--..."
|     <!--   -->
|   <body>

#data
<style>...<!--[if IE]><style>...</style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "...<!--[if IE]><style>..."
|   <body>
|     "X"

#data
<title><!--<title></title>--></title>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <html>
|   <head>
|     <title>
|       "<!--<title>"
|   <body>
|     "-->"

#data
<title>&lt;/title></title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|       "</title>"
|   <body>

#data
<title>foo/title><link></head><body>X
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <title>
|       "foo/title><link></head><body>X"
|   <body>

#data
<noscript><!--<noscript></noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<!--<noscript>"
|   <body>
|     "-->"

#data
<noscript><!--<noscript></noscript>--></noscript>
#errors
 * (1,11) missing DOCTYPE
#script-off
#document
| <html>
|   <head>
|     <noscript>
|       <!-- <noscript></noscript> -->
|   <body>

#data
<noscript><!--</noscript>X<noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<!--"
|   <body>
|     "X"
|     <noscript>
|       "-->"

#data
<noscript><!--</noscript>X<noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#script-off
#document
| <html>
|   <head>
|     <noscript>
|       <!-- </noscript>X<noscript> -->
|   <body>

#data
<noscript><iframe></noscript>X
#errors
(1,10): expected-doctype-but-got-start-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<iframe>"
|   <body>
|     "X"

#data
<noscript><iframe></noscript>X
#errors
 * (1,11) missing DOCTYPE
 * (1,19) unexpected token in head noscript
 * (1,31) unexpected EOF
#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <iframe>
|       "</noscript>X"

#data
<noframes><!--<noframes></noframes>--></noframes>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <html>
|   <head>
|     <noframes>
|       "<!--<noframes>"
|   <body>
|     "-->"

#data
<noframes><body><script><!--...</script></body></noframes></html>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <noframes>
|       "<body><script><!--...</script></body>"
|   <body>

#data
<textarea><!--<textarea></textarea>--></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!--<textarea>"
|     "-->"

#data
<textarea>&lt;/textarea></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "</textarea>"

#data
<iframe><!--<iframe></iframe>--></iframe>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,41): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <iframe>
|       "<!--<iframe>"
|     "-->"

#data
<iframe>...<!--X->...<!--/X->...</iframe>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <iframe>
|       "...<!--X->...<!--/X->..."

#data
<xmp><!--<xmp></xmp>--></xmp>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,29): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <xmp>
|       "<!--<xmp>"
|     "-->"

#data
<noembed><!--<noembed></noembed>--></noembed>
#errors
(1,9): expected-doctype-but-got-start-tag
(1,45): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <noembed>
|       "<!--<noembed>"
|     "-->"

#data
<!doctype html><table>

#errors
(2,0): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "
"

#data
<!doctype html><table><td><span><font></span><span>
#errors
(1,26): unexpected-cell-in-table-body
(1,45): unexpected-end-tag
(1,51): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <span>
|               <font>
|             <font>
|               <span>

#data
<!doctype html><form><table></form><form></table></form>
#errors
(1,35): unexpected-end-tag-implies-table-voodoo
(1,35): unexpected-end-tag
(1,41): unexpected-form-in-table
(1,56): unexpected-end-tag
(1,56): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <form>
|       <table>
|         <form>
